Combatants refactor

+++ xivsupport/src/main/java/gg/xp/events/actlines/data/Job.java
@@ -100,13 +100,13 @@ public enum Job {
 	}
 
 
-	public static Job getById(int id) {
+	public static Job getById(long id) {
 		// This will work as long as they stay in order and contiguous
 		if (id > values().length) {
 			log.error("There is no job with ID {}, using ADV instead", id);
 			return ADV;
 		}
-		return values()[id];
+		return values()[(int) id];
 	}
 
 	public int getId() {
